refactor(report): extract ReportScreen subwidgets and ReportType enum
ReportScreen was 520 LOC. Split it per #571:

- lib/features/report/domain/entities/report_type.dart — ReportType enum
  (all variant routing logic and display names)
- lib/features/report/presentation/widgets/no_backend_banner.dart —
  #484 banner when neither Tankerkoenig nor TankSync is configured
- lib/features/report/presentation/widgets/report_input_section.dart —
  the dynamic price / free-text field below the radio list
- lib/features/report/presentation/widgets/report_type_list.dart —
  the "What's wrong?" header + radio list

Screen file is now 319 LOC (from 520). report_type.dart is re-exported
from report_screen.dart so existing `import '.../report_screen.dart'`
consumers (tests, providers) keep working without churn.

Note: the radio list is intentionally a top-level `buildReportTypeList`
function returning `List<Widget>`, not a ConsumerWidget wrapping a Column.
Wrapping 10 RadioListTiles in a Column inside the screen's ListView broke
scroll-extent math on tall lists — the last radios were unreachable via
`tester.scrollUntilVisible` in existing tests. Spreading directly into the
ListView's children preserves per-radio layout as before.

Closes #571
